Added Print Board Functions
I updated the client.py file to include functions that will allow the program to print the game board based upon the input board size and list of strings, which I am calling the list of "labels". I also added a function to allow the validation of users entered positions.